// src/interfaces/IGlobalVariables.ts
/**
 * Created by Srikanta Dutta on 2/13/2017.
 */

'use strict';
export interface IGlobalVariables {
    viewType: string;
    nodeID: string;
    showSpinner: boolean;
    isGetOperation: boolean;
    processTabViewType: string;
    isShowActivityTab: boolean;
    env: any;
    LoadUAA: boolean,
    space: any;
    applicationsCount: any;
    spaces: any;
    toggleHide: any;
    selectedTabProcessDiagram: string;
    configDefaultContainer: string;
    appLevel2CurrentTab: string;
    hideConfigLeftPanel: boolean;
    configView: string;
}